pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Александр Мышкин <sir.myshkin@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
python-dotenv = "^1.0.0"
dj-database-url = "^2.0.0"
flake8 = "^6.0.0"
gunicorn = "^20.1.0"
whitenoise = "^6.5.0"
psycopg2 = "^2.9.6"
django-bootstrap5 = "^23.3"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
